app/requirements.txt
alembic==1.5.5
aniso8601==9.0.0
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==0.16.1
